修复好友获取不全的问题

+++ items.go
@@ -125,7 +125,7 @@ type WebWxSyncResponse struct {
 
 type WebWxContactResponse struct {
 	MemberCount  int
-	Seq          int
+	Seq          int64
 	BaseResponse BaseResponse
 	MemberList   []*User
 }
